refactoring

+++ Giveaway.SteamGifts/Commands/Driver/DriverRemoveProfileCommand.cs
@@ -1,9 +1,12 @@
 ﻿using Giveaway.SteamGifts.Models;
 
+using NLog;
+
 namespace Giveaway.SteamGifts.Commands
 {
     internal class DriverRemoveProfileCommand : BaseCommand
     {
+        public ILogger Logger => LogManager.GetCurrentClassLogger();
         public string DriverProfilePath { get; set; }
         public DriverRemoveProfileCommand(Configuration configuration) : base(configuration)
         {
@@ -12,9 +15,29 @@ namespace Giveaway.SteamGifts.Commands
 
         public override void Execute()
         {
-            Directory.Delete(DriverProfilePath, true);
-            Console.WriteLine("Профиль ChromeDriver успешно удален...");
-            Console.ReadLine();
+
+            try
+            {
+                if (Directory.Exists(DriverProfilePath))
+                {
+                    Directory.Delete(DriverProfilePath, true);
+                    Console.WriteLine("Профиль ChromeDriver успешно удален...");
+                }
+                else
+                {
+                    Console.WriteLine($"Профиль по пути {Path.Combine(Directory.GetCurrentDirectory(), DriverProfilePath)} отсутствует");
+                }
+            }
+            catch (Exception ex)
+            {
+                Logger.Error(ex);
+                Console.WriteLine("Ошибка во время удаления профиля");
+            }
+            finally
+            {
+                Console.WriteLine("Press any key to continue...");
+                Console.ReadLine();
+            }
         }
     }
 }
